RH: more cleanup
